+++ yash-env/src/lib.rs
@@ -445,6 +445,7 @@ impl Env {
     }
 }
 
+mod alias;
 pub mod builtin;
 pub mod function;
 pub mod input;