+++ yash-env/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.1] - Unreleased
 
+### Added
+
+- `impl yash_system::alias::Glossary for Env`
+
 ### Changed
 
 - External dependency versions: